[Refactor/#54]: follow api 분리, 프로필 수정 api 통일
