--- src/features/Test/api/test.api.ts
@@ -1,9 +1,0 @@
-import { baseOptions } from '@/api/options';
-import { EditProfile, EditProfileReq } from '@frolog/frolog-api';
-
-const editProfile = new EditProfile(baseOptions);
-
-export const editTestType = async (req: EditProfileReq) => {
-  const data = await editProfile.fetch(req);
-  return data.result;
-};
